Fixed support for all buttons. Added left analog stick.
Fixed the support for all buttons by basing on the MegaJoy example. Instead of putting everything in a loop, the bitshifting and everything is handled one by one. Messy but it will work. Analog stick has been added. The Y axis had to be inverted but otherwise works fine. Some more work on the limits may have to be done.